+++ bld/wv/win/h/winscrn.h
@@ -30,7 +30,7 @@
 ****************************************************************************/
 
 
-extern void __far       *ExtraAlloc( unsigned );
+extern void __far       *ExtraAlloc( size_t );
 extern void             ExtraFree( void __far * );
 extern void             win_uisetcolor( int );
 extern void             win_uisetmono( void );
